Add files via upload
Completed requirements: Combining login and dashboard in a single file, commented out separate login and dashboard files, added HOME to navigate back to home page on all screens, inserted username and password to admin db at the start of the app and use that to validate login, added validation for row and column number. This completes all functional requirements given in the project.